requirements.txt
aiofiles==24.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
httpx==0.28.1
idna==3.11
lxml==6.0.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.2.1
sanic==24.6.0
sanic-cors==2.2.0
urllib3==2.6.3
google-cloud-secret-manager==2.21.1
firebase-admin==7.1.0
google-auth==2.47.0
google-auth-oauthlib==1.2.1
sanic-session==0.8.0
